hotel_service/requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-spectacular==0.27.1
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Pillow==10.1.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
